feat: add vault contracts
This commit adds contracts and tests for a segregated vault system for
asset wrappers. The main contracts are as follows:
- VaultFactory - the NFT contract which spawns new vaults upon
`initializeBundle` calls. It mints a new NFT token representing
ownership of the vault and sends it to the creator
- AssetVault - Contract for storing segregated bundles of assets,
including ERC20, ERC721, ERC1155, and ETH. Users can create a vault from
the factory and safely send in their assets to the vault, using it as
collateral in bundle form. When they want to withdraw, they must
explicity `enableWithdraw` on their vault, which changes the state and
allows withdrawals of any deposited assets. Note that once withdrawals
are enabled they cannot be disabled - this is a safety feature for
integrations like the OriginationController, which need to assert that
the assets a potential lender saw in the vault are still there.
