+++ contracts/LoanCore.sol
@@ -9,7 +9,7 @@ import "@openzeppelin/contracts/token/ERC20/IERC20.sol";
 import "@openzeppelin/contracts/token/ERC20/utils/SafeERC20.sol";
 import "@openzeppelin/contracts/utils/Counters.sol";
 import "./interfaces/IPromissoryNote.sol";
-import "./interfaces/IAssetWrapper.sol";
+import "./interfaces/IAssetVault.sol";
 import "./interfaces/IFeeController.sol";
 import "./interfaces/ILoanCore.sol";
 
@@ -30,9 +30,9 @@ contract LoanCore is ILoanCore, AccessControl, Pausable {
     Counters.Counter private loanIdTracker;
     mapping(uint256 => LoanLibrary.LoanData) private loans;
     mapping(uint256 => bool) private collateralInUse;
-    IPromissoryNote public override borrowerNote;
-    IPromissoryNote public override lenderNote;
-    IERC721 public override collateralToken;
+    IPromissoryNote public immutable override borrowerNote;
+    IPromissoryNote public immutable override lenderNote;
+    IERC721 public immutable override collateralToken;
     IFeeController public override feeController;
 
     // 10k bps per whole
@@ -52,8 +52,6 @@ contract LoanCore is ILoanCore, AccessControl, Pausable {
 
         // Avoid having loanId = 0
         loanIdTracker.increment();
-
-        emit Initialized(address(collateralToken), address(borrowerNote), address(lenderNote));
     }
 
     /**
