--- contracts/interfaces/IAssetWrapper.sol
@@ -1,109 +1,0 @@
-// SPDX-License-Identifier: MIT
-
-pragma solidity ^0.8.0;
-
-/**
- * @dev Interface for an AssetWrapper contract
- */
-interface IAssetWrapper {
-    /**
-     * @dev Emitted when an ERC20 token is deposited
-     */
-    event DepositERC20(address indexed depositor, uint256 indexed bundleId, address tokenAddress, uint256 amount);
-
-    /**
-     * @dev Emitted when an ERC721 token is deposited
-     */
-    event DepositERC721(address indexed depositor, uint256 indexed bundleId, address tokenAddress, uint256 tokenId);
-
-    /**
-     * @dev Emitted when an ERC1155 token is deposited
-     */
-    event DepositERC1155(
-        address indexed depositor,
-        uint256 indexed bundleId,
-        address tokenAddress,
-        uint256 tokenId,
-        uint256 amount
-    );
-
-    /**
-     * @dev Emitted when ETH is deposited
-     */
-    event DepositETH(address indexed depositor, uint256 indexed bundleId, uint256 amount);
-
-    /**
-     * @dev Emitted when a bundle is unwrapped.
-     */
-    event Withdraw(address indexed withdrawer, uint256 indexed bundleId);
-
-    /**
-     * @dev Creates a new bundle token for `to`. Its token ID will be
-     * automatically assigned (and available on the emitted {IERC721-Transfer} event)
-     *
-     * See {ERC721-_mint}.
-     */
-    function initializeBundle(address to) external returns (uint256);
-
-    /**
-     * @dev Deposit some ERC20 tokens into a given bundle
-     *
-     * Requirements:
-     *
-     * - The bundle with id `bundleId` must have been initialized with {initializeBundle}
-     * - `amount` tokens from `msg.sender` on `tokenAddress` must have been approved to this contract
-     */
-    function depositERC20(
-        address tokenAddress,
-        uint256 amount,
-        uint256 bundleId
-    ) external;
-
-    /**
-     * @dev Deposit an ERC721 token into a given bundle
-     *
-     * Requirements:
-     *
-     * - The bundle with id `bundleId` must have been initialized with {initializeBundle}
-     * - The `tokenId` NFT from `msg.sender` on `tokenAddress` must have been approved to this contract
-     */
-    function depositERC721(
-        address tokenAddress,
-        uint256 tokenId,
-        uint256 bundleId
-    ) external;
-
-    /**
-     * @dev Deposit an ERC1155 token into a given bundle
-     *
-     * Requirements:
-     *
-     * - The bundle with id `bundleId` must have been initialized with {initializeBundle}
-     * - The `tokenId` from `msg.sender` on `tokenAddress` must have been approved for at least `amount`to this contract
-     */
-    function depositERC1155(
-        address tokenAddress,
-        uint256 tokenId,
-        uint256 amount,
-        uint256 bundleId
-    ) external;
-
-    /**
-     * @dev Deposit some ETH into a given bundle
-     *
-     * Requirements:
-     *
-     * - The bundle with id `bundleId` must have been initialized with {initializeBundle}
-     */
-    function depositETH(uint256 bundleId) external payable;
-
-    /**
-     * @dev Withdraw all assets in the given bundle, returning them to the msg.sender
-     *
-     * Requirements:
-     *
-     * - The bundle with id `bundleId` must have been initialized with {initializeBundle}
-     * - The bundle with id `bundleId` must be owned by or approved to msg.sender
-     */
-    function withdraw(uint256 bundleId) external;
-}
