+++ contracts/test/MockOpenVault.sol
@@ -1,0 +1,7 @@
+pragma solidity ^0.8.0;
+
+contract MockOpenVault {
+    function withdrawEnabled() external pure returns (bool) {
+        return false;
+    }
+}
